Simplify OpFactory
